requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
dj-database-url==2.1.0
google-genai==1.45.0
python-dotenv==1.1.1